+++ src/LuminaSupplemental.Excel/Model/FieldOpCofferType.cs
@@ -1,0 +1,8 @@
+namespace LuminaSupplemental.Excel.Model;
+
+public enum FieldOpCofferType
+{
+    Bronze = 1,
+    Silver = 2,
+    Gold = 3,
+}