+++ src/LuminaSupplemental.Excel/Model/FieldOpType.cs
@@ -1,0 +1,11 @@
+namespace LuminaSupplemental.Excel.Model;
+
+public enum FieldOpType
+{
+    Pagos = 1,
+    Pyros = 2,
+    Hydatos = 3,
+    OccultTreasure = 4,
+    OccultPot = 5,
+    OccultGoldenCoffer = 6,
+}