+++ src/LuminaSupplemental.Excel/Model/ItemSupplement.cs
@@ -19,16 +19,22 @@ namespace LuminaSupplemental.Excel.Model
         [Name("ItemId")] public uint ItemId { get; set; }
         [Name("SourceItemId")] public uint SourceItemId { get; set; }
         [Name("ItemSupplementSource"), TypeConverter(typeof( EnumConverter ))] public ItemSupplementSource ItemSupplementSource { get; set; }
+        [Name("Min")] public uint? Min { get; set; }
+        [Name("Max")] public uint? Max { get; set; }
+        [Name("Probability")] public decimal? Probability { get; set; }
 
         public RowRef< Item > Item;
 
         public RowRef< Item > SourceItem;
 
-        public ItemSupplement(uint itemId, uint sourceItemId, ItemSupplementSource itemSupplementSource )
+        public ItemSupplement(uint itemId, uint sourceItemId,ItemSupplementSource itemSupplementSource,  uint? min = null, uint? max = null, decimal? probability = null )
         {
             ItemId = itemId;
             SourceItemId = sourceItemId;
             ItemSupplementSource = itemSupplementSource;
+            Min = min;
+            Max = max;
+            Probability = probability;
         }
 
         public ItemSupplement()
@@ -43,6 +49,21 @@ namespace LuminaSupplemental.Excel.Model
             if( Enum.TryParse<ItemSupplementSource>( lineData[ 2 ], out var itemSupplementSource ) )
             {
                 ItemSupplementSource = itemSupplementSource;
+            }
+
+            if (lineData[3] != string.Empty)
+            {
+                Min = uint.Parse( lineData[ 3 ], CultureInfo.InvariantCulture );
+            }
+
+            if (lineData[4] != string.Empty)
+            {
+                Max = uint.Parse( lineData[ 4 ], CultureInfo.InvariantCulture );
+            }
+
+            if (lineData[5] != string.Empty)
+            {
+                Probability = decimal.Parse( lineData[ 5 ], CultureInfo.InvariantCulture );
             }
         }