+++ src/LuminaSupplemental.Excel/Model/ItemSupplementSource.cs
@@ -7,6 +7,17 @@ public enum ItemSupplementSource
     Loot = 3,
     Gardening = 4,
     SkybuilderHandIn = 5,
+    CardPacks = 6,
+    Coffer = 7,
+    PalaceOfTheDead = 8,
+    HeavenOnHigh = 9,
+    EurekaOrthos = 10,
+    Anemos = 11,
+    Pagos = 12,
+    Pyros = 13,
+    Hydatos = 14,
+    Bozja = 15,
+    Logogram = 16,
     Unknown = 999
-    
+
 }